% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lrm}
\alias{lrm}
\title{LRM}
\usage{
lrm(
  y,
  x,
  test.variable = 1,
  test.matrix = matrix(c(0)),
  intercept = TRUE,
  estimate = TRUE,
  alpha = 0.05,
  interpretation = FALSE,
  anova = FALSE,
  Ftest = FALSE,
  partialtest = FALSE,
  confidence.interval = FALSE,
  GLH = FALSE
)
}
\arguments{
\item{y}{the dependent variable}

\item{x}{the independent invariable}

\item{test.variable}{the chosen parameter to be tested, default is the first term}

\item{test.matrix}{contrast matrix for GLH, default is an empty matrix}

\item{intercept}{"TRUE" means model include the intercept term, "FALSE" means not}

\item{estimate}{"TRUE" means calculate the point estimation of parameters, "FALSE" means not}

\item{alpha}{the significance level, default is 0.05}

\item{interpretation}{"TRUE" means print the result of hypothesis testing, "FALSE" means not}

\item{anova}{"TRUE" means print anova table, "FALSE" means not}

\item{Ftest}{"TRUE" means conduct F test, "FALSE" means not}

\item{partialtest}{"TRUE" means conduct partial test, "FALSE" means not}

\item{confidence.interval}{"TRUE" means print the confidence interval, "FALSE" means not}

\item{GLH}{"TRUE" means conduct GLH, "FLASE" means not}
}
\value{
result of one specific hypothesis testing
}
\description{
This package is used to fit a linear regression model to inference and predict
}
\examples{
lrm(mtcars$mpg, mtcars$wt)
#'lrm(mtcars$mpg, mtcars$wt, estimate=FALSE, anova=TRUE, Ftest=TRUE)
lrm(mtcars$mpg, mtcars$wt, test.variable=2, estimate=FALSE, partialtest=TRUE)
lrm(mtcars$mpg, mtcars$wt, test.matrix=matrix(c(0,1),nrow=1,ncol=2), estimate=FALSE, GLH=TRUE)


}
